feat: organize code into layers
- Service layer: core business logic. App is a collection of services. Services can optionally have underlying implementors (e.g. Firebase as an auth service).
- Repo ports: business logic around persistence
- Repo adapters: handle read/write of primitive (json) data in e.g. postgres.
- Router ports: business logic for handling incoming requests.
- Router adapters (e.g. axum): Parse raw requests into technology agnostic requests.

Routing adapter -> Router -> Service -> Repo -> Repo adapter.

App is a trait/struct that contains the service layer. Each service has it's own repo (if necessary). Routers are top level functions because there's nowhere to keep an instance. The global-ish app instance is the one given to Axum as the app's state. That's how services are called upon when handling a request.